requirements.txt
Appium-Flutter-Finder==0.5.0
Appium-Python-Client==2.11.1
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
decorator==5.1.1
dnspython==2.4.2
docutils==0.20.1
exceptiongroup==1.1.3
h11==0.14.0
idna==3.4
kitchen==1.2.6
outcome==1.2.0
pycparser==2.21
pymongo==4.5.0
PySocks==1.7.1
robotframework==6.1.1
robotframework-appiumflutterlibrary==1.0.0
robotframework-appiumlibrary==2.0.0
selenium==4.12.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.10.4
urllib3==2.0.4
wsproto==1.2.0
